requirements.txt
pandas==1.5.0
dash==2.6.2
pyserial==3.5
